% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assembler.R
\name{assembler}
\alias{assembler}
\title{start community assembly process.}
\usage{
assembler(community, control = NULL, filename = NULL, prev = NULL)
}
\arguments{
\item{community}{A \code{community} object.}

\item{control}{A control object affecting how the assembly proceeds,
see \code{assembler_control} for details,}

\item{filename}{Location to save results (currently does nothing).}

\item{prev}{History from previously run assembly, used to restart.}
}
\value{
An \code{assembler} object.
}
\description{
Initialise a community assembly object
}
\details{
Control options affect how the assembly proceeds.
}
\author{
Rich FitzJohn, Daniel Falster
}

