% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assembler.R
\name{assembler_control}
\alias{assembler_control}
\title{Options controllings community assembly process.}
\usage{
assembler_control(control = NULL)
}
\arguments{
\item{control}{A list of values to modify from defaults.}
}
\value{
A list with elements run_type, birth_type, birth_move_tol,
compute_viable_fitness, n_mutants, n_immigrants, check_positive,
vcv, check_inviable, dead_seed_rain, eps_too_close
}
\description{
Controls how community assembly works.
}
\details{
Returns a list. Passing in a list of value via \code{
control} will override the defaults. Options include
run_type determines whether population is stepped to
demographic equilibrium ("to_equilibrium") or not ("single").
"birth_type" determines sampling of new types -- "stochastic" or
"maximum" (on fitness peak). With "stochastic" births,
"n_mutants" and "n_immigrants" determine the frequency of
resident mutations and immigrations from global pool.
"vcv" is variance-covariance matrix for mutations.
If "birth_move_tol" is trait distance
within which we attempt to move an existing resident rather introduce
a new type (this helps reduce the number of types).
"compute_viable_fitness" asks whether to check bounds of viable
trait space. "check_positive" determines whether the fitness of an
invader is checked before it is introduced. If
"check_inviable" causes dead residents to be removed when seed rain
drops below "dead_seed_rain".
"eps_too_close" is tolerance in trait values when searching for maxima.
}
\author{
Rich FitzJohn, Daniel Falster
}

