% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembler.R
\name{assembler_run}
\alias{assembler_run}
\title{Take specified number of steps with assembler.}
\usage{
assembler_run(obj, nsteps)
}
\arguments{
\item{obj}{An  \code{assembler} object.}

\item{nsteps}{Number of steps}
}
\value{
An \code{assembler} object.
}
\description{
Take specified number of steps with assembler.
}
\author{
Rich FitzJohn, Daniel Falster
}
