% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bird_log_fitness}
\alias{bird_log_fitness}
\title{Bird model: log invasion fitness of mutants}
\usage{
bird_log_fitness(x_mut, x_res, n_res, pars)
}
\arguments{
\item{x_mut}{numeric vector of mutant trait values (arrival times)}

\item{x_res}{numeric vector of resident trait values}

\item{n_res}{numeric vector of resident equilibrium densities}

\item{pars}{list with a, x_opt, sigma, R0, K, p}
}
\value{
numeric vector of log invasion fitness, one per mutant
}
\description{
Bird model: log invasion fitness of mutants
}
\keyword{internal}
