% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_demography.R
\name{community_demography}
\alias{community_demography}
\title{Update demography of community}
\usage{
community_demography(community)
}
\arguments{
\item{community}{A \code{community} object.}
}
\value{
A \code{community} object.
}
\description{
Update demography of community according to specified rules
}
\author{
Rich FitzJohn, Daniel Falster
}
