% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_fitness_landscape.R
\name{community_fitness_landscape}
\alias{community_fitness_landscape}
\title{Fitness Landscape}
\usage{
community_fitness_landscape(
  community,
  method = community$fitness_control$method,
  ...
)
}
\arguments{
\item{community}{A community object}

\item{method}{used to construct landscape}

\item{...}{additional arguments passed to the chosen landscape method
(e.g. \code{bounds}, \code{n_evals}).}
}
\description{
Construct a fitness landscape.
}
\author{
Daniel Falster
}
