% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_fitness.R
\name{community_fitness_landscape_bayes}
\alias{community_fitness_landscape_bayes}
\title{Construct a fitness landscape using Bayesian optimisation.}
\usage{
community_fitness_landscape_bayes(
  community,
  bounds = community$bounds,
  npts = community$fitness_control$npts,
  ninit = community$fitness_control$ninit
)
}
\description{
Construct a fitness landscape using Bayesian optimisation.
}
