% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_fitness.R
\name{community_fitness_landscape_bayesopt}
\alias{community_fitness_landscape_bayesopt}
\title{Construct a fitness landscape using Bayesian optimisation.}
\usage{
community_fitness_landscape_bayesopt(
  community,
  bounds = community$bounds,
  n_evals = community$fitness_control$n_evals,
  n_init = community$fitness_control$n_init
)
}
\description{
Construct a fitness landscape using Bayesian optimisation.
}
