% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_attractors.R
\name{community_selection_gradient_1d}
\alias{community_selection_gradient_1d}
\title{Calculates selection gradient in a single-species community
with given trait value}
\usage{
community_selection_gradient_1d(community, dx = 1e-04, log_scale = TRUE)
}
\arguments{
\item{community}{community object to use.}

\item{dx}{Interval over which derivative is calculated}

\item{log_scale}{(currently disabled) Determines whether derivative is taken
with respect to raw or log-transformed x values. The latter
is useful when x is log-normally distributed.}
}
\value{
a community with selection gradient added.
}
\description{
Adds selection gradient to a single-species community
with given trait value. This is derivative of fitness
with respect to trait value. You should first solv for
using \code{community_run_to_equilibrium}
}
\author{
Daniel Falster
}
