% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_attractors.R
\name{community_solve_singularity_1D}
\alias{community_solve_singularity_1D}
\title{Find evolutionary attractor for single species and trait.}
\usage{
community_solve_singularity_1D(
  community,
  bounds = NULL,
  tol = 1e-04,
  ...,
  edge_ok = TRUE
)
}
\arguments{
\item{bounds}{a vector containing the end-points of the
interval to be searched for the root.}

\item{tol}{the desired accuracy (convergence tolerance).}

\item{...}{set verbpse=TRUE for verbose output, birth_rate=value
gives starting values when solving for demographic equilibrium}

\item{edge_ok}{Is it (not) an error if we end up on the edge of the
viable bounds?}

\item{trait}{name of trait}

\item{p}{Parameters object to use.  Importantly, the
\code{strategy_default} element gets used here.}
}
\value{
A species object, with trait, seed rain and cohort schedule
information.
}
\description{
Find evolutionary attractor for single species and trait.
This is point at which selection gradient equals zero.
Currently solved using \code{uniroot}
}
\author{
Daniel Falster
}
