% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{community_start}
\alias{community_start}
\title{Initialise a community object}
\usage{
community_start(
  bounds,
  birth_rate_initial = 0.001,
  demography_control = demographic_step_control(),
  fitness_control = NULL,
  model_support = NULL,
  harness = NULL,
  trait_scale = c("log", "linear")
)
}
\arguments{
\item{bounds}{A set of bounds, as specified in \code{plant}.}

\item{birth_rate_initial}{A vector of birth rates.}

\item{demography_control}{List controlling the demography/equilibrium
solver. See \code{\link{demographic_step_control}}.}

\item{fitness_control}{List of parameters controlling
how approximate fitness landscapes are generated. See function
\code{fitness_control} for an example.}

\item{model_support}{Model support data passed to the harness (e.g. the
\code{plant} parameters and control objects).}

\item{harness}{A model harness wiring the community to a demographic
backend (see \code{\link{harness_plant}}); defaults to the \code{plant} SCM.}

\item{trait_scale}{How trait space is spaced/searched during assembly:
\code{"log"} for strictly positive traits or \code{"linear"} for traits
spanning zero.}
}
\value{
An \code{community} object.
}
\description{
Initialise a community object
}
\details{
Used to store full description of community. Is a list
with elements parameters, bounds, birth_rate_initial,
trait_names, traits, birth_rate, fitness_control.
}
\author{
Rich FitzJohn, Daniel Falster
}
