% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dd99_equilibrium}
\alias{dd99_equilibrium}
\title{DD99 model: resident demographic equilibrium densities}
\usage{
dd99_equilibrium(x_res, pars)
}
\arguments{
\item{x_res}{numeric vector of resident trait values}

\item{pars}{list with r, K0, x0, sigma_K, sigma_C}
}
\value{
numeric vector of equilibrium densities
}
\description{
Single resident: N* = K(x). Many residents: solve the linear system A N = K
with A_ij = C(x_i - x_j) (negative densities, i.e. strategies that cannot
coexist, are clamped to zero).
}
\keyword{internal}
