% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dd99_fitness}
\alias{dd99_fitness}
\title{DD99 model: invasion fitness of mutants (a per-capita growth rate)}
\usage{
dd99_fitness(x_mut, x_res, n_res, pars)
}
\arguments{
\item{x_mut}{numeric vector of mutant trait values}

\item{x_res}{numeric vector of resident trait values}

\item{n_res}{numeric vector of resident equilibrium densities}

\item{pars}{list with r, K0, x0, sigma_K, sigma_C}
}
\value{
numeric vector of invasion fitness (=0 for a resident at equilibrium)
}
\description{
DD99 model: invasion fitness of mutants (a per-capita growth rate)
}
\keyword{internal}
