% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dd99_nd_equilibrium}
\alias{dd99_nd_equilibrium}
\title{DD99 model (nD): resident demographic equilibrium densities}
\usage{
dd99_nd_equilibrium(x_res, pars)
}
\arguments{
\item{x_res}{numeric matrix of resident trait values}

\item{pars}{list with r, K0, x0, sigma_K, sigma_C (the last three length-k)}
}
\value{
numeric vector of equilibrium densities
}
\description{
DD99 model (nD): resident demographic equilibrium densities
}
\keyword{internal}
