% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_approximate.R
\name{fitness_approximate_control}
\alias{fitness_approximate_control}
\title{Control how fitness landscapes are constructed}
\usage{
fitness_approximate_control(control = NULL)
}
\arguments{
\item{control}{List of values to change from default.}
}
\value{
List with elements type, n, finite_only,
n_initial,, n_each, n_predict, lower_limit, cost,
x_seed.
}
\description{
List of parameters controlling
how approximate fitness landscapes are generated.
}
\author{
Rich FitzJohn, Daniel Falster
}
