% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/python.R
\name{for_python_equilibrium}
\alias{for_python_equilibrium}
\alias{for_python_evolve}
\alias{for_python_fitness}
\alias{for_python_fitness_approximate}
\alias{for_python_fitness_prepare}
\alias{for_python_simulate}
\title{Wrapper code for Python}
\usage{
for_python_equilibrium(time_disturbance, slope, lma, seed_rain,
  equilibrium_nsteps = 20L, verbose = FALSE)

for_python_evolve(time_disturbance, slope, nsteps, lma = NULL,
  seed_rain = NULL, verbose = FALSE)

for_python_simulate(time_disturbance, slope, traits, nsteps = 100)

for_python_fitness(sys, lma)

for_python_fitness_approximate(sys, lma)

for_python_fitness_prepare(sys)
}
\arguments{
\item{time_disturbance}{The mean time to disturbance}

\item{slope}{Indicates strength of a leaf economic spectrum tradeoff}

\item{lma}{Vector of species traits}

\item{seed_rain}{Vector of species abundances}

\item{equilibrium_nsteps}{Maximum number of steps to run the
equilibrium rain for}

\item{verbose}{Be noisy during calculation of equilibrium seed
rain.}

\item{nsteps}{Number of evolutionary steps to run for}

\item{sys}{Result of running \code{for_python} or
\code{for_python_evolve}}

\item{equilibrium}{Logical, inidicating if we should compute
equilibrium seed rain (the default) or not.}
}
\description{
This is a small helper function for calling tree from Python.  As
we develop the ideas better, this will change.  The current verison
allows computing of the fitness landscape for a set of parameters,
similar to the test case we've been talking about.
}

