% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{geritz99_log_fitness}
\alias{geritz99_log_fitness}
\title{Geritz 1999 seed-size model: log invasion fitness of mutants}
\usage{
geritz99_log_fitness(x_mut, x_res, n_res, pars)
}
\arguments{
\item{x_mut}{numeric vector of mutant seed sizes}

\item{x_res}{numeric vector of resident seed sizes}

\item{n_res}{numeric vector of resident equilibrium densities (seeds/site)}

\item{pars}{list with R, alpha, beta}
}
\value{
numeric vector of log invasion fitness (=0 for resident at equilibrium;
  -Inf for non-viable mutants where s(x) = 0)
}
\description{
Geritz 1999 seed-size model: log invasion fitness of mutants
}
\keyword{internal}
