% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{geritz_log_fitness}
\alias{geritz_log_fitness}
\title{Geritz 1998 soft-selection model: log invasion fitness of mutants}
\usage{
geritz_log_fitness(x_mut, x_res, n_res, pars)
}
\arguments{
\item{x_mut}{numeric vector of mutant trait values}

\item{x_res}{numeric vector of resident trait values}

\item{n_res}{numeric vector of resident equilibrium abundances}

\item{pars}{list with sigma, mu (patch optima), K (patch capacities)}
}
\value{
numeric vector of log invasion fitness (=0 for resident at equilibrium)
}
\description{
Geritz 1998 soft-selection model: log invasion fitness of mutants
}
\keyword{internal}
