% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harness.R
\name{harness_dd99}
\alias{harness_dd99}
\title{DD99 (Dieckmann & Doebeli 1999) harness}
\usage{
harness_dd99(
  r = 1,
  K0 = 500,
  x0 = 0,
  sigma_K = 1,
  sigma_C = 0.4,
  trait_name = "x"
)
}
\arguments{
\item{r}{intrinsic growth rate}

\item{K0}{maximum carrying capacity}

\item{x0}{trait optimising the resource (singular strategy)}

\item{sigma_K}{width of the resource/carrying-capacity kernel}

\item{sigma_C}{width of the competition kernel}

\item{trait_name}{name of the evolving trait}
}
\value{
a `harness` object
}
\description{
DD99: Dieckmann & Doebeli 1999 competition model (Nature 400:354-357).
}
\details{
Continuous-time logistic competition for a Gaussian resource: carrying
capacity \code{K(x) = K0 exp(-(x-x0)^2/(2 sigma_K^2))} and competition kernel
\code{C(d) = exp(-d^2/(2 sigma_C^2))}. Invasion fitness
\code{s(y) = r (1 - sum_i N_i C(y-x_i)/K(y))}. The singular strategy is
\code{x* = x0}; it is an evolutionary \emph{branching point} iff
\code{sigma_C < sigma_K} and an ESS iff \code{sigma_C > sigma_K} -- the
classic disruptive-selection oracle.
}
\author{
Daniel Falster
}
