% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assembler.R
\name{length.assembler}
\alias{length.assembler}
\title{Returns number taken by an \code{assembler} object}
\usage{
\method{length}{assembler}(x)
}
\arguments{
\item{x}{An  \code{assembler} object.}
}
\value{
Length of history in assembler object.
}
\description{
Returns number taken by an assembler object
}
\author{
Rich FitzJohn, Daniel Falster
}

