% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/community.R
\name{length.community}
\alias{length.community}
\title{Returns number of types in \code{community}}
\usage{
\method{length}{community}(x)
}
\arguments{
\item{x}{A \code{community} object.}
}
\value{
number of rows in trait matrix.
}
\description{
Returns number of types in community
}
\author{
Rich FitzJohn, Daniel Falster
}

