% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/local_landscape.R
\name{local_landscape}
\alias{local_landscape}
\title{Generate local landscape}
\usage{
local_landscape(sys, xlim = NULL, ylim = NULL, scal = 0.25, n = 101,
  combine = "closest", traits = NULL)
}
\arguments{
\item{sys}{Single community with element
\code{fitness_approximate_slopes}}

\item{xlim,ylim}{Optional x and y limits (otherwise resident range
will be expanded by \code{scal})}

\item{scal}{Scaling parameter when creating \code{xlim} and
\code{ylim}.}

\item{n}{Number of points to compute fitness for (in each of x and
y, so \code{n^2} points total are used)}

\item{combine}{Leave this be.}

\item{traits}{Which traits to compute the matrix for.  Not
required with two traits, but for more than two traits, you must
indicate which pair of traits to compute the landscape for.}
}
\description{
Generate a local landscape from a history with a
\code{fitness_approximate_slopes} element.
}

