% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_plant.R
\name{plant_community_check_for_inviable_strategies}
\alias{plant_community_check_for_inviable_strategies}
\title{Check low-abundance strategies for viability}
\usage{
plant_community_check_for_inviable_strategies(community)
}
\arguments{
\item{community}{A community object}
}
\description{
Check low-abundance strategies for viability.
}
