% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_plots.R
\name{plot_community}
\alias{plot_community}
\title{Returns birth rates for residents in community}
\usage{
plot_community(tidy_community, ...)
}
\arguments{
\item{tidy_community}{Community or history object from tidied assembly}

\item{...}{additional arguments passed to \code{plot_community_1d}.}
}
\value{
Returns one plot if community of, if history, plots length of history timeseries
}
\description{
Returns birth rates for residents in community
}
