% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_plots.R
\name{plot_community_2d}
\alias{plot_community_2d}
\title{Plot pair-wise trait combinations of resident community}
\usage{
plot_community_2d(tidy_community, ...)
}
\arguments{
\item{tidy_community}{Community or history object from tidied assembly with two traits}

\item{...}{}
}
\value{
Returns one plot if community of, if history, plots length of history timeseries
}
\description{
Plot pair-wise trait combinations of resident community
}
