% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_plots.R
\name{plot_landscape}
\alias{plot_landscape}
\title{Plot fitness landscape}
\usage{
plot_landscape(tidy_landscape, ...)
}
\arguments{
\item{tidy_landscape}{Landscape object from tidied assembly}

\item{...}{}
}
\value{
community_landscape returns one plot, history_landscape returns plots length of timeseries
}
\description{
Plot fitness landscape
}
