% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_output.R
\name{tidy_assembly}
\alias{tidy_assembly}
\title{Tidies output from assembler_run into tidy dataframe.}
\usage{
tidy_assembly(obj)
}
\arguments{
\item{obj}{ouytput of an assembly run}
}
\value{
A tibble of the assembly
}
\description{
Tidies output from assembler_run into tidy dataframe.
}
