% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_plant.R
\name{community_fitness_landscape}
\alias{community_fitness_landscape}
\title{Fitness Landscape}
\usage{
community_fitness_landscape(
  community,
  bounds = community$bounds,
  npts = community$fitness_control$n
)
}
\arguments{
\item{community}{A community object}

\item{bounds}{a bounds object}

\item{npts}{number of points}
}
\description{
Construct a fitness landscape.
}
\author{
Daniel Falster
}
