% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{community_start}
\alias{community_start}
\title{Initialise a community object}
\usage{
community_start(
  parameters,
  bounds,
  birth_rate_initial = 0.001,
  hyperpar = NULL,
  fitness_control = NULL
)
}
\arguments{
\item{parameters}{A \code{parameters} object, as specified
in \code{plant}..}

\item{bounds}{A set of bounds, as specified in \code{plant}.}

\item{birth_rate_initial}{A vector of birth rates.}

\item{hyperpar}{A plant hyperparameter function to be used when calling \code{strategy_list}}

\item{fitness_control}{List of parameters controlling
how approximate fitness landscapes are generated. See function
\code{fitness_control} for an example.}
}
\value{
An \code{community} object.
}
\description{
Initialise a community object
}
\details{
Used to store full description of community. Is a list
with elements parameters, bounds, birth_rate_initial,
trait_names, traits, birth_rate, fitness_control.
}
\author{
Rich FitzJohn, Daniel Falster
}
