% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_landscape.R
\name{local_landscape}
\alias{local_landscape}
\alias{local_landscape_control}
\title{Generate local landscape}
\usage{
local_landscape(sys, xlim = NULL, ylim = NULL, control = NULL, traits = NULL)

local_landscape_control(control = NULL)
}
\arguments{
\item{sys}{Single community with element
\code{fitness_slopes}}

\item{xlim, ylim}{Optional x and y limits (otherwise resident range
will be expanded by \code{control$scal})}

\item{control}{Set of control parameters, passed to
\code{local_landscape_control}}

\item{traits}{Which traits to compute the matrix for.  Not
required with two traits, but for more than two traits, you must
indicate which pair of traits to compute the landscape for.}
}
\description{
Generate a local landscape from a history with a
\code{fitness_slopes} element.
}
