% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/births_stochastic.R
\name{mutational_vcv_proportion}
\alias{mutational_vcv_proportion}
\title{Generate mutational covariance matrices}
\usage{
mutational_vcv_proportion(x, p = 0.001)
}
\arguments{
\item{x}{A community object, or appropriate matrix of bounds}

\item{p}{Fraction of trait range to mutate (vectors will be
recycled according R's usual rules).}
}
\description{
Generate a mutational variance covariance matrix where each trait
mutates by a fraction of its total range, and where there is no
covariance among traits.  This is a helper function for use with
\code{\link{assembler_stochastic_naive}}
}

