% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_plots.R
\name{community_plot_fitness_landscape}
\alias{community_plot_fitness_landscape}
\title{Plot fitness landscape}
\usage{
community_plot_fitness_landscape(
  community,
  step = NA,
  xlim = community$bounds,
  ylim = c(-20, 20)
)
}
\arguments{
\item{community}{XXXX}

\item{...}{}
}
\value{
community_landscape returns one plot, history_landscape returns plots length of timeseries
}
\description{
Plot fitness landscape
}
