% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_output.R
\name{tidy_assembly}
\alias{tidy_assembly}
\title{Tidies output from assembler_run.}
\usage{
tidy_assembly(obj)
}
\arguments{
\item{assembly}{}
}
\value{
Returns the birth rate of the final community as well as the historical timeseries of birth rates for residents. If fitness landscapes are available, provides those as well.
}
\description{
Tidies output from assembler_run.
}
